+++ src/background/background.ts
