+++ src/options/page.html
@@ -5,14 +5,14 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Options - TTV LOL</title>
-    <link rel="stylesheet" href="../css/boilerplate.css" />
-    <link rel="stylesheet" href="../css/options.css" />
+    <link rel="stylesheet" href="../common/css/boilerplate.css" />
+    <link rel="stylesheet" href="./style.css" />
   </head>
   <body>
     <header>
       <div class="logo">
-        <img src="../../images/icon.png" alt="Icon of TTV LOL" />
-        <img src="../img/ttvlol.png" alt="TTV LOL" />
+        <img src="../assets/icon.png" alt="Icon of TTV LOL" />
+        <img src="../assets/logo.png" alt="TTV LOL" />
       </div>
     </header>
     <main>
@@ -67,6 +67,6 @@
       </section>
     </main>
 
-    <script type="module" src="../ts/options.ts"></script>
+    <script type="module" src="./options.ts"></script>
   </body>
 </html>
