+++ src/popup/menu.html
@@ -1,6 +1,6 @@
 <html>
   <head>
-    <link rel="stylesheet" href="../css/styles.css" />
+    <link rel="stylesheet" href="./style.css" />
     <meta content="text/html;charset=utf-8" http-equiv="Content-Type" />
     <meta content="utf-8" http-equiv="encoding" />
   </head>
@@ -8,7 +8,7 @@
   <body>
     <div class="container">
       <div style="background-color: #202127; border-radius: 8px; margin: 30px">
-        <img width="60%" src="../img/ttvlol.png" />
+        <img width="60%" src="../assets/logo.png" />
       </div>
       <div class="button-group">
         <a
